requirements.txt
numpy>=1.20.0
scipy>=1.7.0
opencv-python>=4.5.0
matplotlib>=3.3.0
imageio>=2.9.0
Pillow>=8.0.0

